Tidied up some of the label code.
Now slightly darker so easier to see, but the scramble is still visible underneath, still looks worse in multiBLD though, just because of the size of the scramble images.
Also changed 3x3x3 label code to be more consistent with the others, and also so the label is printed after the rest of the face is drawn.